+++ lib/snmp_table_viewer/fetcher.rb
@@ -7,7 +7,7 @@ module SNMPTableViewer
     # @param base_oid [String] The OID to start the SNMP walk from
     # @param **snmp_options [Hash] The options to pass to NETSNMP::Client.new
     # @return [Array<Array<#to_s>>] A two dimensional array containing objects in each cell (at 'address' data[row][col])
-    def self.fetch(base_oid:, **snmp_options)
+    def self.from_snmp(base_oid:, **snmp_options)
       data = Array.new
       NETSNMP::Client.new(snmp_options) do |manager|
         manager.walk(oid: base_oid).each do |oid, value|