+++ lib/snmp_table_viewer/formatter/raw.rb
@@ -1,7 +1,10 @@
 module SNMPTableViewer
 
+  # Formatting class for raw output.
   class Formatter::Raw < Formatter
-    def output(**options)
+    # Output the data (and headings if provided).
+    # @return [String] the raw data
+    def output()
       data_with_headings.inspect
     end
   end # Raw Fromatter