+++ spec/snmp_table_viewer/fetcher_spec.rb
@@ -1,9 +1,5 @@
 describe SNMPTableViewer::Fetcher do
 
-  it 'Create' do
-    expect(described_class.new()).to_not be_nil
-  end
-
   it 'Fetches' do
     snmp_options = {opt_a: :a}
     manager = double("manager")
@@ -18,8 +14,7 @@ describe SNMPTableViewer::Fetcher do
     expect(NETSNMP::Client).to receive(:new).with(snmp_options).once do |&block|
       block.call(manager)
     end # expect block
-    fetcher = described_class.new(snmp_options)
-    expect(fetcher.fetch('1.2.3.4.5')).to eq [
+    expect(described_class.from_snmp(base_oid: '1.2.3.4.5', **snmp_options)).to eq [
       [1, 'a', 'A'],
       [2, 'b', 'B']
     ]