+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Version 0.0.2
 
-  * Fetcher.fetch is now a class mehtod
+  * Fetcher.fetch is now a class mehtod called from_snmp
 
 ## Version 0.0.1