+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Version 0.0.2
+
+  * Fetcher.fetch is now a class mehtod called from_snmp
+
 ## Version 0.0.1
 
   * Gemification of a personal utility